After implementing #14, I've been wondering if that could cause memory issues. I can think of some memory leaks caused by deliberate behavior, but I also think that in most cases it should not cause user-perceivable problems.

Still, I don't like that approach, saving the value on each link makes me feel insecure. Finally I discovered that we can have a 
push-pull-push model solution that doesn't use version/value at all.

The new push-pull-push model is very similar to the approach adopted by https://github.com/vuejs/core/pull/5912.

- First push: signal/computed pushes dirty/toCheckDirty flag to subscribers
- Second pull: computed/effect pull updates from toCheckDirty dependencies when executed
- Final push: When computed is pulled, if the value is updated, pushes dirty flat to shallow subscribers

In this method, we don't need to use intermediate variables to detect changes, because all flag updates are immediate as needed.

### Memory Usage improvement

Since Link does not need to store value, the memory usage is reduced by 3% (2434.72KB -> 2363.90KB) in the following test.

```
const start = process.memoryUsage().heapUsed;

const w = 100;
const h = 100;
const src = signal(1);

for (let i = 0; i < w; i++) {
	let last = src;
	for (let j = 0; j < h; j++) {
		const prev = last;
		last = computed(() => prev.get() + 1);
	}
	effect(() => last.get());
}

src.set(src.get() + 1);

console.log(`Memory Usage: ${((process.memoryUsage().heapUsed - start) / 1024).toFixed(2)} KB`);
```

### Benchmark Results

#### main bench

| benchmark            |              avg |         min |         p75 |         p99 |         max |
| -------------------- | ---------------- | ----------- | ----------- | ----------- | ----------- |
| propagate: 1 * 1     | `  1.43 µs/iter` | `  1.42 µs` | `  1.43 µs` | `  1.45 µs` | `  1.45 µs` |
| propagate: 1 * 10    | `  7.40 µs/iter` | `  7.38 µs` | `  7.41 µs` | `  7.42 µs` | `  7.42 µs` |
| propagate: 1 * 100   | ` 65.68 µs/iter` | ` 64.79 µs` | ` 65.58 µs` | ` 69.79 µs` | `101.50 µs` |
| propagate: 10 * 1    | ` 12.85 µs/iter` | ` 12.84 µs` | ` 12.85 µs` | ` 12.85 µs` | ` 12.85 µs` |
| propagate: 10 * 10   | ` 72.23 µs/iter` | ` 71.38 µs` | ` 72.13 µs` | ` 75.88 µs` | `104.50 µs` |
| propagate: 10 * 100  | `653.37 µs/iter` | `649.33 µs` | `654.33 µs` | `673.13 µs` | `703.71 µs` |
| propagate: 100 * 1   | `126.44 µs/iter` | `125.21 µs` | `126.17 µs` | `134.21 µs` | `156.04 µs` |
| propagate: 100 * 10  | `717.02 µs/iter` | `712.42 µs` | `717.71 µs` | `741.50 µs` | `765.29 µs` |
| propagate: 100 * 100 | `  6.51 ms/iter` | `  6.49 ms` | `  6.51 ms` | `  6.60 ms` | `  6.61 ms` |
| complex: 1 * 1       | `  1.79 µs/iter` | `  1.76 µs` | `  1.78 µs` | `  2.45 µs` | `  2.98 µs` |
| complex: 1 * 10      | ` 10.64 µs/iter` | ` 10.57 µs` | ` 10.60 µs` | ` 10.88 µs` | ` 10.94 µs` |
| complex: 1 * 100     | `103.02 µs/iter` | `101.42 µs` | `102.83 µs` | `113.33 µs` | `384.17 µs` |
| complex: 10 * 1      | ` 16.39 µs/iter` | ` 16.23 µs` | ` 16.30 µs` | ` 16.74 µs` | ` 16.76 µs` |
| complex: 10 * 10     | `109.98 µs/iter` | `108.25 µs` | `109.79 µs` | `116.75 µs` | `362.38 µs` |
| complex: 10 * 100    | `  1.08 ms/iter` | `  1.08 ms` | `  1.08 ms` | `  1.12 ms` | `  1.27 ms` |
| complex: 100 * 1     | `165.35 µs/iter` | `163.42 µs` | `165.08 µs` | `179.42 µs` | `209.79 µs` |
| complex: 100 * 10    | `  1.12 ms/iter` | `  1.11 ms` | `  1.12 ms` | `  1.16 ms` | `  1.31 ms` |
| complex: 100 * 100   | ` 12.63 ms/iter` | ` 12.33 ms` | ` 12.54 ms` | ` 14.26 ms` | ` 14.50 ms` |

#### This PR

| benchmark            |              avg |         min |         p75 |         p99 |         max |
| -------------------- | ---------------- | ----------- | ----------- | ----------- | ----------- |
| propagate: 1 * 1     | `  1.43 µs/iter` | `  1.42 µs` | `  1.43 µs` | `  1.44 µs` | `  1.47 µs` |
| propagate: 1 * 10    | `  7.17 µs/iter` | `  7.14 µs` | `  7.18 µs` | `  7.19 µs` | `  7.20 µs` |
| propagate: 1 * 100   | ` 63.22 µs/iter` | ` 63.15 µs` | ` 63.25 µs` | ` 63.26 µs` | ` 63.31 µs` |
| propagate: 10 * 1    | ` 12.78 µs/iter` | ` 12.77 µs` | ` 12.79 µs` | ` 12.79 µs` | ` 12.79 µs` |
| propagate: 10 * 10   | ` 70.18 µs/iter` | ` 69.38 µs` | ` 70.00 µs` | ` 75.04 µs` | ` 95.21 µs` |
| propagate: 10 * 100  | `640.52 µs/iter` | `630.67 µs` | `644.63 µs` | `653.75 µs` | `674.08 µs` |
| propagate: 100 * 1   | `125.57 µs/iter` | `124.13 µs` | `125.33 µs` | `136.38 µs` | `155.38 µs` |
| propagate: 100 * 10  | `697.68 µs/iter` | `693.50 µs` | `698.25 µs` | `720.67 µs` | `755.00 µs` |
| propagate: 100 * 100 | `  6.32 ms/iter` | `  6.30 ms` | `  6.32 ms` | `  6.38 ms` | `  6.44 ms` |
| complex: 1 * 1       | `  1.78 µs/iter` | `  1.75 µs` | `  1.76 µs` | `  2.51 µs` | `  2.98 µs` |
| complex: 1 * 10      | ` 10.58 µs/iter` | ` 10.50 µs` | ` 10.53 µs` | ` 10.77 µs` | ` 10.81 µs` |
| complex: 1 * 100     | `100.13 µs/iter` | ` 98.54 µs` | ` 99.92 µs` | `106.38 µs` | `386.96 µs` |
| complex: 10 * 1      | ` 16.34 µs/iter` | ` 16.15 µs` | ` 16.59 µs` | ` 16.71 µs` | ` 16.72 µs` |
| complex: 10 * 10     | `106.88 µs/iter` | `104.96 µs` | `106.67 µs` | `114.96 µs` | `391.38 µs` |
| complex: 10 * 100    | `  1.05 ms/iter` | `  1.05 ms` | `  1.05 ms` | `  1.08 ms` | `  1.24 ms` |
| complex: 100 * 1     | `162.52 µs/iter` | `160.17 µs` | `163.21 µs` | `175.88 µs` | `201.54 µs` |
| complex: 100 * 10    | `  1.10 ms/iter` | `  1.09 ms` | `  1.10 ms` | `  1.15 ms` | `  1.26 ms` |
| complex: 100 * 100   | ` 12.37 ms/iter` | ` 12.10 ms` | ` 12.27 ms` | ` 13.96 ms` | ` 14.27 ms` |